icpac-booking-backend/requirements.txt
Django==5.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.2.0
celery==5.3.4
redis==5.0.1
Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0